Refactor travel modules into cartographer travel package
