+++ PluginOverview.txt
@@ -23,7 +23,8 @@ src/
 │  ├─ map-gallery.ts         # Historische Galerie; leitet jetzt direkt in den Cartographer um
 │  ├─ map-editor/            # Interaktiver Hex-Editor – MapEditorOverview.txt
 │  ├─ terrain-editor/        # Terrain-Palette – TerrainEditorOverview.txt
-│  ├─ travel-guide/          # Legacy-View, mountet den Cartographer-Travel-Mode – TravelGuideOverview.txt
+│  ├─ legacy/
+│  │  └─ travel-guide-view/  # Legacy-View, mountet den Cartographer-Travel-Mode
 │  └─ cartographer/          # Multi-Mode Map Viewer (Travel/Editor/Inspector) – CartographerOverview.txt
 │
 ├─ core/                     # Hex-Engine & Dateiservices – CoreOverview.txt
@@ -44,7 +45,7 @@ src/
 - **Terrain Editor (`terrain-editor/`):** Pflegt Farb- und Geschwindigkeits-Paletten über den Core-Terrain-Store und synchronisiert Änderungen mit `terrain.ts`.
 - **Map Gallery (`map-gallery.ts`):** Historische View, die beim direkten Öffnen (z. B. aus alten Workspaces) weiterhin auf den Cartographer umleitet.
 - **Map Editor (`map-editor/`):** Liefert Tooling für Hex-Auswahl, Brush-Painting, Tile-Speicherung und Inspector-Views (Details siehe `MapEditorOverview.txt`); wird über neue Cartographer-Flows eröffnet.
-- **Travel Guide (`travel-guide/`):** Legacy-Hülle für Rückwärtskompatibilität. Mountet `createTravelGuideMode()` in derselben `sm-cartographer`-Shell, verwaltet nur View-Lifecycle und delegiert sämtliche UI/Logik an den Cartographer.
+- **Legacy Travel Guide (`legacy/travel-guide-view/`):** Dünne Hülle für Rückwärtskompatibilität. Mountet den Travel-Modus aus `cartographer/travel/` in der bekannten `sm-cartographer`-Shell und kümmert sich ausschließlich um den View-Lifecycle.
 
 ### Core-Services (`src/core/`)
 - Bündelt Hex-Mathematik, SVG-Rendering, Map/Terrain-Dateiverwaltung und Workspace-Utilities.
