+++ src/apps/cartographer/travel/render/draw-route.ts
